refactor: rename to ddb package
