+++ server/ddb/transactions.go
@@ -1,4 +1,4 @@
-package dynamodb
+package ddb
 
 import (
 	"github.com/aws/aws-sdk-go/service/dynamodb"
