+++ server/ddb/users_test.go
@@ -1,4 +1,4 @@
-package dynamodb
+package ddb
 
 import (
 	"testing"
